+++ quilter.h
@@ -1,9 +1,9 @@
 //
-//  quilter.h
-//  quilter
+//	quilter.h
+//	quilter
 //
-//  Created by Jeffrey Lomicka on 5/1/25.
-//  Copyright © 2025 Jeffrey Lomicka. All rights reserved.
+//	Created by Jeffrey Lomicka on 5/1/25.
+//	Copyright © 2025 Jeffrey Lomicka. All rights reserved.
 //
 
 #ifndef quilter_h
@@ -19,11 +19,11 @@ public:
 	bool iReadyToExecute = false;				// Set this to true when you want Execute to run.
 	bool iNeedNewPrompt = false;				// If you printf anything, you should request a new command line prompt
 	char iDescription[ 512];					// Everyone needs one
-    int iExecuteCounter = 0;                    // Number of times Execute has been called
-    char iTimeStringMemory[ 32];
+	int iExecuteCounter = 0;					// Number of times Execute has been called
+	char iTimeStringMemory[ 32];
 
-    virtual void DisplayPrompt();  // This is a hack until I make a base class for command lines
-    int result = 0;                // When set to 2, it means we're exiting. This is also a hack until I make a base class for command lines
+	virtual void DisplayPrompt();  // This is a hack until I make a base class for command lines
+	int result = 0;				   // When set to 2, it means we're exiting. This is also a hack until I make a base class for command lines
 	virtual void Startup() = 0;
 	virtual void Execute() = 0;
 	virtual void Shutdown() = 0;
@@ -33,18 +33,18 @@ public:
 	}
 	virtual ~AsyncHelper();
 #if MACCODE
-    int Printf( const char* formatString, ...) __printflike(2, 3);
+	int Printf( const char* formatString, ...) __printflike(2, 3);
 #else
 	int Printf(const char* formatString, ...);
 #endif
-    char* NowString();
+	char* NowString();
 };
 
 // Utility functions for quilters - some of these might want to move into ConsoleThings or xraise?
 
 size_t ReadFile(FILE *fp, char **buf);
-void AsyncGetLine( char* buffer, size_t len, FILE* f, JeffSemaphore &completionSema, bool* ready);
-void FGetLine( char* buffer, size_t len, FILE* f);
+void AsyncGetLine( char* buffer, int len, FILE* f, JeffSemaphore &completionSema, bool* ready);
+void FGetLine( char* buffer, int len, FILE* f);
 void AddActivity( AsyncHelper* newActivity);	// Adds to list of asynchronous activities
 std::string DocumentsPath();
 std::string SettingsPath();
